AlgoFlow_Architect/requirements.txt
dash==2.14.2
dash-bootstrap-components==1.5.0
google-generativeai
python-dotenv==1.0.0
graphviz==0.20.1
pandas